\name{EEstackWithoutMeta2}
\alias{EEstackWithoutMeta2}
\title{EEstackWithoutMeta2}

\usage{EEstackWithoutMeta2(tarPath, 
    bands = NULL, sat.nm = "LO08")}
\arguments{
  \item{tarPath}{
}
  \item{bands}{
}
  \item{sat.nm}{
}
}



\author{Victor Gutierrez-Velez <victorhugo@temple.edu>, Wilson Lara <wilarhen@temple.edu>, Jeronimo Rodriguez <jeronimo.rodriguez@temple.edu >}





